Refactor graqhql queries using Resolvers
