+++ app/graphql/types/base_field.rb
@@ -3,13 +3,5 @@
 module Types
   class BaseField < GraphQL::Schema::Field
     argument_class Types::BaseArgument
-
-    DATE_FORMAT = /\d{4}-\d{2}-\d{2}/.freeze
-
-    def format_date(date:)
-      return if date.blank?
-
-      date.match(DATE_FORMAT)&.to_s || raise(G3ySchema::InvalidQueryError)
-    end
   end
 end
