+++ app/graphql/types/query_type.rb
@@ -9,32 +9,11 @@ module Types
       argument :id, ID, required: false
       argument :name, String, required: false
     end
-
     field :post, PostType, null: false do
       argument :id, ID, required: true
     end
-
-    field :posts, [PostType], null: false do
-      argument :contentOr, String, required: false, as: :content_or
-      argument :contentAnd, String, required: false, as: :content_and
-      argument :userId, ID, required: false, as: :user_id
-      argument :username, String, required: false
-      argument :createdAtAfter, String,
-               required: false, as: :created_at_after,
-               prepare: ->(date, _ctx) { format_date(date: date) }
-      argument :createdAtBefore, String,
-               required: false, as: :created_at_before,
-               prepare: ->(date, _ctx) { format_date(date: date) }
-      argument :page, Integer, required: false
-      argument :pagePer, Integer, required: false, as: :page_per
-    end
-
-    # field :fuzzy_posts, resolver: Queries::FuzzyPostsQuery
-    field :fuzzy_posts, [PostType], null: false do
-      argument :keyword, String, required: false
-      argument :page, Integer, required: false
-      argument :pagePer, Integer, required: false, as: :page_per
-    end
+    field :posts, resolver: Queries::PostsQuery
+    field :fuzzy_posts, resolver: Queries::FuzzyPostsQuery
 
     def user(id: nil, name: nil)
       User.find(id)
@@ -46,33 +25,7 @@ module Types
       Post.find(id)
     end
 
-    def posts( # rubocop:disable Metrics/ParameterLists
-      content_or: nil, content_and: nil, user_id: nil, username: nil,
-      created_at_after: nil, created_at_before: nil, page: 0, page_per: 20
-    )
-      user = User.find_by(id: user_id) || User.find_by(name: username)
-
-      return [] if (user_id.present? || username.present?) && user.nil?
-
-      Elasticsearch::PostsQuery.new(
-        content_or: content_or, content_and: content_and, user_id: user&.id,
-        created_at_after: created_at_after, created_at_before: created_at_before,
-        page: page, page_per: page_per
-      ).call
-    end
-
     # TODO
     def users(name:); end
-
-    def fuzzy_posts(keyword:, page:, page_per:)
-      # TODO: where or elasticseach
-      # TODO: OR search
-      user = User.find_by(name: keyword)
-
-      Elasticsearch::PostsQuery.new(
-        content_or: keyword, user_id: user&.id,
-        page: page, page_per: page_per
-      ).call
-    end
   end
 end
